#770: Extract field initialization logic from FluentInjector
